tech-review-qacicd/requirements.txt
bandit==1.7.4
boto==2.49.0
boto3==1.24.21
botocore==1.27.21
branca==0.5.0
celery==5.2.7
dateparser==1.1.1
descartes==1.1.0
Django==4.0.5
django-admin-sortable2==2.1
django-choices==1.7.2
django-cors-headers==3.13.0
django-jsonfield2==15.06.08
django-localflavor==3.1
django-multiselectfield==0.1.12
django-redis==5.2.0
django-ses==3.1.0
django-storages==1.12.3
django-treebeard==4.5.1
django-user-agents==0.4.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
django-multiselectfield==0.1.12
drf-yasg==1.20.0
flower==1.0.0
fpdf==1.7.2
geopandas==0.11.0
geojson==2.5.0
google-cloud-storage==2.4.0
importlib-metadata<5.0.0
ipython==8.4.0
jedi==0.18.1
Jinja2==2.11.0
matplotlib
markupsafe==2.0.1
numba
numpy>=1.23
pandas==1.5.0
pdfminer.six==20220524
Pillow==9.1.1
pre-commit==2.19.0
plotly==5.9.0
psutil==5.9.1
psycopg2-binary==2.9.3
pyarrow==9.0.0
pygeos==0.12.0 --no-binary=pygeos
PyMuPDF==1.20.1
pytz==2022.1
qrcode==7.3.1
rasterio==1.2.10
redis==4.3.4
requests==2.28.1
requests_oauthlib==1.3.1
rtree==1.0.0
selenium==4.3.0
sentry-sdk==1.6.0
Shapely==1.8.5 --no-binary=Shapely
scikit-learn==1.1.1
uWSGI==2.0.20
watchtower==3.0.0
weasyprint==51
debugpy==1.5.1